requirements.txt
flask==2.0.1
kivy==2.1.0
python-telegram-bot==13.7
python-dotenv==0.19.0
SpeechRecognition==3.8.1
pyaudio==0.2.11
asyncio==3.4.3 